+++ src/city.cpp
@@ -636,7 +636,7 @@ inline Terrain *getTerrainAt(City *city, s32 x, s32 y)
 	return result;
 }
 
-void drawTerrain(City *city, Renderer *renderer, Rect2I visibleArea, s32 shaderID)
+void drawTerrain(City *city, Renderer *renderer, Rect2I visibleArea, s8 shaderID)
 {
 	DEBUG_FUNCTION_T(DCDT_GameUpdate);
 
@@ -742,7 +742,7 @@ Building* getBuildingAtPosition(City *city, s32 x, s32 y)
 	return result;
 }
 
-void drawBuildings(City *city, Renderer *renderer, Rect2I visibleTileBounds, s32 shaderID, Rect2I demolitionRect)
+void drawBuildings(City *city, Renderer *renderer, Rect2I visibleTileBounds, s8 shaderID, Rect2I demolitionRect)
 {
 	DEBUG_FUNCTION_T(DCDT_GameUpdate);
